init commit with some scripts to scrape letterbox
Letterbox contains lots of good user curated "lists" we can use these as training and testing data if we do go down the prompt to slate approach. The script currently only goes to 30 movies for each list but we could expand that as a 30 movie list may cause a smaller hit rate when it might still be a hit.